+++ carl/circuit.py
@@ -1,6 +1,6 @@
 import shapely.geometry as geom
 from descartes import PolygonPatch
-
+import numpy as np
 
 class Circuit(object):
 
@@ -98,3 +98,16 @@ class Circuit(object):
         ax.set_xlim(bounds[0] - offset_x, bounds[2] + offset_x)
         ax.set_ylim(bounds[1] - offset_y, bounds[3] + offset_y)
         ax.set_aspect(1)
+
+
+def generateCircuitPoints(n_points=16, difficulty=0, circuit_size=(5, 2)):
+    n_points = min(25, n_points)
+    angles = [-np.pi/4 + 2*np.pi*k/n_points for k in range(3*n_points//4)]
+    points = [(circuit_size[0]/2, 0.5), (3*circuit_size[0]/2, 0.5)]
+    points += [(circuit_size[0]*(1+np.cos(angle)), circuit_size[1]*(1+np.sin(angle))) for angle in angles]
+    for i, angle in zip(range(n_points), angles):
+        rd_dist = 0
+        if difficulty > 0:
+            rd_dist = min(circuit_size) * np.random.vonmises(mu=0, kappa=32/difficulty)/np.pi
+        points[i+2] = tuple(np.array(points[i+2]) + rd_dist*np.array([np.cos(angle), np.sin(angle)]))
+    return points
