+++ requirements.txt
@@ -1,8 +1,8 @@
-shapely
-Keras
 tensorflow
-numpy
+shapely
+numpy==1.19.3
 descartes
 jupyter
-h5py
+h5py<3.0.0
 matplotlib
+gym
